opam: depend on ssl package required for HTTPS
Install the ssl package to avoid the error "No SSL or TLS support
compiled into Conduit" when downloading from HTTPS. The other OPAM
package that we could use instead of ssl is tls. However, when we tried
to depend on tls, the ocaml version was restricted, and we encountered
compilation errors with other packages for some reason.

Signed-off-by: Iglói Gábor <gabor.igloi@citrix.com>